// api-gateway/src/modules/users/infrastructure/proto/com/ecommerceapp/v1/users.ts
export enum Gender {
  GENDER_UNSPECIFIED = 0,
  GENDER_MALE = 1,
  GENDER_FEMALE = 2,
  GENDER_OTHER = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum Role {
  ROLE_UNSPECIFIED = 0,
  ROLE_CUSTOMER = 1,
  ROLE_SHIPPER = 2,
  ROLE_SHOPOWNER = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface UserId {
  userId: string;
}

export interface CreateUserRequest {
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  username: string;
  password: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  nation: string;
  gender: Gender;
  role: Role;
  dateOfBirth:
    | Date
    | undefined;
  /** Optional fields for one-to-one relationships */
  customer?: CreateCustomerCommand | undefined;
  shipper?: CreateShipperCommand | undefined;
  shopOwner?: CreateCustomerCommand | undefined;
}

/** Define fields for Customer here if needed */
export interface CreateCustomerCommand {
}

/** Define fields for ShopOwner here if needed */
export interface CreateShopOwnerCommand {
}

export interface CreateShipperCommand {
  /** Define fields for Shipper here if needed */
  shipmentInventoryId: string;
}
// ...
export interface UpdateUserInfoRequest {
  firstName?: string | undefined;
  lastName?: string | undefined;
  phoneNumber?: string | undefined;
  addressLine1?: string | undefined;
  addressLine2?: string | undefined;
  city?: string | undefined;
  nation?: string | undefined;
  gender?: Gender | undefined;
  role?: Role | undefined;
  dateOfBirth?: Date | undefined;
  customer?: UpdateCustomerRequest | undefined;
  shipper?: UpdateShipperRequest | undefined;
  shopOwner?: UpdateShopOwnerRequest | undefined;
}

export interface UpdateCustomerRequest {
}

export interface UpdateShipperRequest {
  shipmentInventoryId?: string | undefined;
}

export interface UpdateShopOwnerRequest {
}

export interface UpdateUserRequest {
  userId: UserId | undefined;
  updateInfo: UpdateUserInfoRequest | undefined;
}

export interface UpdateUserResponse {
  id: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  addressLine1: string;
  addressLine2?: string | undefined;
  city: string;
  nation: string;
  gender: Gender;
  role: Role;
  dateOfBirth: Date | undefined;
  customer?: Customer | undefined;
  shipper?: Shipper | undefined;
  shopOwner?: ShopOwner | undefined;
}

export interface ChangePasswordRequest {
  userId: string;
  oldPassword: string;
  newPassword: string;
}

export interface UpdateRefreshTokenRequest {
  userId: string;
  refreshToken: string;
}

export interface ValidateUserRequest {
  username: string;
  password: string;
}

export interface Empty {
}

function createBaseUserId(): UserId {
  return { userId: "" };
}

export const UserId: MessageFns<UserId> = {
  encode(message: UserId, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== "") {
      writer.uint32(10).string(message.userId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UserId {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUserId();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UserId {
    return { userId: isSet(object.userId) ? globalThis.String(object.userId) : "" };
  },

  toJSON(message: UserId): unknown {
    const obj: any = {};
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UserId>, I>>(base?: I): UserId {
    return UserId.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UserId>, I>>(object: I): UserId {
    const message = createBaseUserId();
    message.userId = object.userId ?? "";
    return message;
  },
};
// ...
export const CreateCustomerCommand: MessageFns<CreateCustomerCommand> = {
  encode(_: CreateCustomerCommand, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateCustomerCommand {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateCustomerCommand();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): CreateCustomerCommand {
    return {};
  },

  toJSON(_: CreateCustomerCommand): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateCustomerCommand>, I>>(base?: I): CreateCustomerCommand {
    return CreateCustomerCommand.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateCustomerCommand>, I>>(_: I): CreateCustomerCommand {
    const message = createBaseCreateCustomerCommand();
    return message;
  },
};
// ...
export const CreateShipperCommand: MessageFns<CreateShipperCommand> = {
  encode(message: CreateShipperCommand, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.shipmentInventoryId !== "") {
      writer.uint32(10).string(message.shipmentInventoryId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateShipperCommand {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateShipperCommand();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.shipmentInventoryId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateShipperCommand {
    return {
      shipmentInventoryId: isSet(object.shipmentInventoryId) ? globalThis.String(object.shipmentInventoryId) : "",
    };
  },

  toJSON(message: CreateShipperCommand): unknown {
    const obj: any = {};
    if (message.shipmentInventoryId !== "") {
      obj.shipmentInventoryId = message.shipmentInventoryId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateShipperCommand>, I>>(base?: I): CreateShipperCommand {
    return CreateShipperCommand.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateShipperCommand>, I>>(object: I): CreateShipperCommand {
    const message = createBaseCreateShipperCommand();
    message.shipmentInventoryId = object.shipmentInventoryId ?? "";
    return message;
  },
};
// ...
export const UpdateUserRequest: MessageFns<UpdateUserRequest> = {
  encode(message: UpdateUserRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.userId !== undefined) {
      UserId.encode(message.userId, writer.uint32(10).fork()).join();
    }
    if (message.updateInfo !== undefined) {
      UpdateUserInfoRequest.encode(message.updateInfo, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): UpdateUserRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUpdateUserRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.userId = UserId.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.updateInfo = UpdateUserInfoRequest.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): UpdateUserRequest {
    return {
      userId: isSet(object.userId) ? UserId.fromJSON(object.userId) : undefined,
      updateInfo: isSet(object.updateInfo) ? UpdateUserInfoRequest.fromJSON(object.updateInfo) : undefined,
    };
  },

  toJSON(message: UpdateUserRequest): unknown {
    const obj: any = {};
    if (message.userId !== undefined) {
      obj.userId = UserId.toJSON(message.userId);
    }
    if (message.updateInfo !== undefined) {
      obj.updateInfo = UpdateUserInfoRequest.toJSON(message.updateInfo);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<UpdateUserRequest>, I>>(base?: I): UpdateUserRequest {
    return UpdateUserRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UpdateUserRequest>, I>>(object: I): UpdateUserRequest {
    const message = createBaseUpdateUserRequest();
    message.userId = (object.userId !== undefined && object.userId !== null)
      ? UserId.fromPartial(object.userId)
      : undefined;
    message.updateInfo = (object.updateInfo !== undefined && object.updateInfo !== null)
      ? UpdateUserInfoRequest.fromPartial(object.updateInfo)
      : undefined;
    return message;
  },
};
// ...
export interface UsersServiceClient extends Client {
  createUser(
    request: CreateUserRequest,
    callback: (error: ServiceError | null, response: CreateUserResponse) => void,
  ): ClientUnaryCall;
  createUser(
    request: CreateUserRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: CreateUserResponse) => void,
  ): ClientUnaryCall;
  createUser(
    request: CreateUserRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: CreateUserResponse) => void,
  ): ClientUnaryCall;
  updateUser(
    request: UpdateUserRequest,
    callback: (error: ServiceError | null, response: UpdateUserResponse) => void,
  ): ClientUnaryCall;
  updateUser(
    request: UpdateUserRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: UpdateUserResponse) => void,
  ): ClientUnaryCall;
  updateUser(
    request: UpdateUserRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: UpdateUserResponse) => void,
  ): ClientUnaryCall;
  changePassword(
    request: ChangePasswordRequest,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  changePassword(
    request: ChangePasswordRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  changePassword(
    request: ChangePasswordRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  updateRefreshToken(
    request: UpdateRefreshTokenRequest,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  updateRefreshToken(
    request: UpdateRefreshTokenRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  updateRefreshToken(
    request: UpdateRefreshTokenRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  validateUser(
    request: ValidateUserRequest,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  validateUser(
    request: ValidateUserRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
  validateUser(
    request: ValidateUserRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: Empty) => void,
  ): ClientUnaryCall;
}

export const UsersServiceClient = makeGenericClientConstructor(
  UsersServiceService,
  "com.ecommerceapp.v1.users.UsersService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): UsersServiceClient;
  service: typeof UsersServiceService;
  serviceName: string;
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };